+++ src/test/java/com/hileco/cortex/instructions/conditions/EQUALSTest.kt
@@ -18,7 +18,7 @@ class EQUALSTest : InstructionTest() {
                 .headingParagraph("EQUALS").paragraph("The EQUALS operation removes two elements from the stack, then adds a 1 or 0 to the stack" + " depending on whether the top element was equal to the second element.")
                 .paragraph("Example program:").source(instructions)
                 .paragraph("Resulting stack:").source(stack)
-        Assert.assertEquals(stack.size().toLong(), 1)
+        Assert.assertEquals(stack.size(), 1)
         Assert.assertArrayEquals(stack.pop(), ConditionInstruction.TRUE)
     }
 }
