+++ src/test/java/com/hileco/cortex/instructions/jumps/JUMP_IFTest.kt
@@ -22,6 +22,6 @@ class JUMP_IFTest : InstructionTest() {
                         "JUMP_DESTINATION instruction. The JUMP_DESTINATION by itself is equal to a NOOP.")
                 .paragraph("Example program:").source(instructions)
                 .paragraph("Resulting stack:").source(stack)
-        Assert.assertEquals(stack.size().toLong(), 0)
+        Assert.assertEquals(stack.size(), 0)
     }
 }
